Add snapshot support

+++ Lab.AkkaNet.Banking.Actors/EventSourcedExample/Account.cs
@@ -10,14 +10,14 @@ namespace Lab.AkkaNet.Banking.Actors.EventSourcedExample
     public class Account : EventSourcedUntypedActor
     {
 
-        public static Props Create(int number, double initialBalance) => Props.Create(() => new Account(number, initialBalance));
+        public static Props Create(int number, decimal initialBalance) => Props.Create(() => new Account(number, initialBalance));
 
         private int number;
 
-        private double balance;
+        private decimal balance;
 
 
-        public Account(int number, double initialBalance)
+        public Account(int number, decimal initialBalance)
         {
             this.number = number;
             this.balance = initialBalance;
@@ -64,7 +64,7 @@ namespace Lab.AkkaNet.Banking.Actors.EventSourcedExample
 
     public class Deposit
     {
-        public Deposit(Guid transactionId, int number, double amount)
+        public Deposit(Guid transactionId, int number, decimal amount)
         {
             TransactionId = transactionId;
             Number = number;
@@ -73,13 +73,13 @@ namespace Lab.AkkaNet.Banking.Actors.EventSourcedExample
 
         public Guid TransactionId { get; }
         public int Number { get; set; }
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
 
     }
 
     public class AmountDeposited : IEvent
     {
-        public AmountDeposited(Guid transactionId, int number, double amount)
+        public AmountDeposited(Guid transactionId, int number, decimal amount)
         {
             TransactionId = transactionId;
             Number = number;
@@ -88,14 +88,14 @@ namespace Lab.AkkaNet.Banking.Actors.EventSourcedExample
 
         public Guid TransactionId { get; }
         public int Number { get; set; }
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
 
     }
 
     public class Withdraw
     {
 
-        public Withdraw(Guid transactionId, int number, double amount)
+        public Withdraw(Guid transactionId, int number, decimal amount)
         {
             TransactionId = transactionId;
             Number = number;
@@ -104,14 +104,14 @@ namespace Lab.AkkaNet.Banking.Actors.EventSourcedExample
 
         public Guid TransactionId { get; }
         public int Number { get;  }
-        public double Amount { get;  }
+        public decimal Amount { get;  }
 
     }
 
     public class AmountWithdrawn : IEvent
     {
 
-        public AmountWithdrawn(Guid transactionId, int number, double amount)
+        public AmountWithdrawn(Guid transactionId, int number, decimal amount)
         {
             TransactionId = transactionId;
             Number = number;
@@ -120,7 +120,7 @@ namespace Lab.AkkaNet.Banking.Actors.EventSourcedExample
 
         public Guid TransactionId { get; }
         public int Number { get;  }
-        public double Amount { get;  }
+        public decimal Amount { get;  }
 
     }
 }
