+++ Lab.AkkaNet.Banking.Actors/UntypedExample/Bank.cs
@@ -39,7 +39,7 @@ namespace Lab.AkkaNet.Banking.Actors.UntypedExample
                     //account.Tell(new QueryBalance(queryAccountBalance.Number), Self);
                     //account.Tell(new QueryBalance(queryAccountBalance.Number), Sender);
 
-                    //var amount = account.Ask<double>(new QueryBalance(queryAccountBalance.Number)).Result;
+                    //var amount = account.Ask<decimal>(new QueryBalance(queryAccountBalance.Number)).Result;
                     break;
             }
         }
@@ -62,7 +62,7 @@ namespace Lab.AkkaNet.Banking.Actors.UntypedExample
     public class Transfer
     {
 
-        public Transfer(int sourceAccountNumber, int targetAccountNumber, double amount)
+        public Transfer(int sourceAccountNumber, int targetAccountNumber, decimal amount)
         {
             SourceAccountNumber = sourceAccountNumber;
             TargetAccountNumber = targetAccountNumber;
@@ -70,20 +70,20 @@ namespace Lab.AkkaNet.Banking.Actors.UntypedExample
         }
         public int SourceAccountNumber { get; }
         public int TargetAccountNumber { get; }
-        public double Amount { get; }
+        public decimal Amount { get; }
     }
 
     public class Open
     {
         
-        public Open(int number, double initialBalance)
+        public Open(int number, decimal initialBalance)
         {
             Number = number;
             InitialBalance = initialBalance;
         }
 
         public int Number { get; }
-        public double InitialBalance { get; }
+        public decimal InitialBalance { get; }
 
     }
 }
